LPD-50383 Add a util class
