+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/constants/CMSSpaceStickerConstants.java
@@ -14,6 +14,4 @@ public class CMSSpaceStickerConstants {
 
 	public static final String SM = "sm";
 
-	public static final String XL = "xl";
-
 }
