+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/SpaceStickerComponentSectionFragmentRenderer.java
@@ -10,6 +10,7 @@ import com.liferay.fragment.renderer.FragmentRendererContext;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.site.cms.site.initializer.internal.constants.CMSSpaceStickerConstants;
 import com.liferay.site.cms.site.initializer.internal.display.context.SpaceStickerDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.util.InfoItemUtil;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -48,7 +49,7 @@ public class SpaceStickerComponentSectionFragmentRenderer
 
 		SpaceStickerDisplayContext spaceStickerDisplayContext =
 			new SpaceStickerDisplayContext(
-				getGroupId(httpServletRequest), _groupLocalService,
+				InfoItemUtil.getGroupId(httpServletRequest), _groupLocalService,
 				httpServletRequest, CMSSpaceStickerConstants.LG);
 
 		return spaceStickerDisplayContext.getProps();
