+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewSpaceContentsAbstractJSPSectionFragmentRenderer.java
@@ -13,6 +13,7 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.site.cms.site.initializer.internal.display.context.ViewSpaceContentsAbstractSectionDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.util.InfoItemUtil;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -37,10 +38,10 @@ public class ViewSpaceContentsAbstractJSPSectionFragmentRenderer
 		HttpServletRequest httpServletRequest) {
 
 		return new ViewSpaceContentsAbstractSectionDisplayContext(
-			_depotEntryLocalService, getGroupId(httpServletRequest),
-			_groupLocalService, httpServletRequest, _language,
-			_objectDefinitionService, _objectDefinitionSettingLocalService,
-			_portal);
+			_depotEntryLocalService,
+			InfoItemUtil.getGroupId(httpServletRequest), _groupLocalService,
+			httpServletRequest, _language, _objectDefinitionService,
+			_objectDefinitionSettingLocalService, _portal);
 	}
 
 	@Override
