+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewSpaceFilesJSPSectionFragmentRenderer.java
@@ -13,6 +13,7 @@ import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.site.cms.site.initializer.internal.display.context.ViewSpaceFilesSectionDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.util.InfoItemUtil;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -37,10 +38,10 @@ public class ViewSpaceFilesJSPSectionFragmentRenderer
 		HttpServletRequest httpServletRequest) {
 
 		return new ViewSpaceFilesSectionDisplayContext(
-			_depotEntryLocalService, getGroupId(httpServletRequest),
-			_groupLocalService, httpServletRequest, _language,
-			_objectDefinitionService, _objectDefinitionSettingLocalService,
-			_portal);
+			_depotEntryLocalService,
+			InfoItemUtil.getGroupId(httpServletRequest), _groupLocalService,
+			httpServletRequest, _language, _objectDefinitionService,
+			_objectDefinitionSettingLocalService, _portal);
 	}
 
 	@Override
